+++ src/index.ts
@@ -1,4 +1,4 @@
 export { default as query } from "./query";
 export { default as schema } from "./schema";
 export { default as createSchemaContext } from "./create-schema-context";
-export { default as syncGraphql } from "./sync-graphql";
+export { default as syncGraphql, isGraphQlPub } from "./sync-graphql";